Add automatic webhook validation
This commit contains an automatic service that will generate a webhook secret for the site that has been created. The secret is used to validate webhook requests from Ghost CMS to GhostMetrics to make sure that the site is allowed to update entities in GhostMetrics. The IntegrationDetail entity has been extended with a Webhook Secret that is automatically generated by the class itself.